move legacy teambinding search to teambinding dir
